inference/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow==2.19.0
pillow==11.2.1
python-multipart==0.0.6
numpy==2.0.2